some refacto
